requirements.txt
streamlit
groq
python-dotenv
chromadb
sentence-transformers
pymupdf
requests
gtts
plotly